@@ -1,21 +1,22 @@
 {
+  "name": "233rooms",
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "scripts": {
-    "migrate": "node-pg-migrate",
     "build": "remix build",
-    "dev": "npm-run-all build --parallel \"dev:*\"",
-    "dev:node": "cross-env NODE_ENV=development nodemon --require dotenv/config ./server.js --watch ./server.js",
-    "dev:remix": "remix watch",
-    "start": "cross-env NODE_ENV=production node ./server.js",
+    "dev": "remix dev --manual",
+    "start": "remix-serve ./build/index.js",
     "typecheck": "tsc"
   },
   "dependencies": {
     "@heroicons/react": "^2.0.17",
     "@react-google-maps/api": "^2.18.1",
-    "@remix-run/express": "^1.15.0",
-    "@remix-run/node": "^1.15.0",
-    "@remix-run/react": "^1.15.0",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@remix-run/css-bundle": "^2.1.0",
+    "@remix-run/node": "^2.1.0",
+    "@remix-run/react": "^2.1.0",
+    "@remix-run/serve": "^2.1.0",
     "@types/googlemaps": "^3.43.3",
     "@uppy/core": "^3.2.1",
     "@uppy/file-input": "^3.0.2",
@@ -32,7 +33,7 @@
     "express": "^4.18.2",
     "framer-motion": "^10.12.4",
     "heroicons": "^2.0.18",
-    "isbot": "^3.6.5",
+    "isbot": "^3.6.8",
     "morgan": "^1.10.0",
     "node-pg-migrate": "^6.2.2",
     "pg": "^8.10.0",
@@ -42,27 +43,30 @@
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-google-places-autocomplete": "^3.4.0",
-    "react-icons": "^4.9.0",
+    "react-icons": "^4.11.0",
     "react-modal": "^3.16.1",
     "react-redux": "^8.0.5",
     "react-swipeable": "^7.0.1",
     "redux": "^4.2.1",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-logger": "^3.0.6",
+    "redux-promise": "^0.6.0",
+    "redux-promise-middleware": "^6.1.3",
     "redux-thunk": "^2.4.2"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.15.0",
-    "@remix-run/eslint-config": "^1.15.0",
-    "@types/bcrypt": "^5.0.0",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.8",
+    "@remix-run/dev": "^2.1.0",
+    "@remix-run/eslint-config": "^2.1.0",
+    "@types/bcrypt": "^5.0.1",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
     "@types/react-modal": "^3.16.0",
-    "dotenv": "^16.0.3",
-    "eslint": "^8.27.0",
-    "nodemon": "^2.0.20",
-    "npm-run-all": "^4.1.5",
-    "typescript": "^4.8.4"
+    "@types/redux-logger": "^3.0.11",
+    "@types/redux-promise": "^0.5.31",
+    "eslint": "^8.38.0",
+    "typescript": "^5.1.6"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18.0.0"
   }
 }
