@@ -7,8 +7,8 @@
 //   user: process.env.DATABASE_USER,
 //   password: '',
 // });
-const { Pool } = require('pg');
-require('dotenv').config();
+import pkg from 'pg';
+const { Pool } = pkg;
 
 export const pool = new Pool({
   connectionString: process.env.POSTGRES_URL,
